requirements.txt
streamlit
transformers
mistralai
Pillow
deepl
